skip if else and
