--- packages/node-type-resolver/tests/PerNodeTypeResolver/VariableTypeResolver/Fixture/AnotherType.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\VariableTypeResolver\Fixture;
-
-final class AnotherType
-{
-
-}
