--- packages/node-type-resolver/tests/PerNodeTypeResolver/VariableTypeResolver/Fixture/ArgumentTypehint.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace SomeNamespace;
-
-use Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\VariableTypeResolver\Fixture\AnotherType;
-
-array_map(function (AnotherType $useUse) {
-    return $useUse;
-}, []);
