+++ packages/node-type-resolver/tests/PerNodeTypeResolver/VariableTypeResolver/Fixture/assignment_class.php.inc
@@ -4,6 +4,10 @@ declare(strict_types=1);
 
 namespace Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\VariableTypeResolver\Fixture;
 
+use Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\VariableTypeResolver\Source\AnotherType;
+use Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\VariableTypeResolver\Source\FirstType;
+use Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\VariableTypeResolver\Source\SecondType;
+
 final class AssignmentClass
 {
     /**
@@ -13,7 +17,7 @@ final class AssignmentClass
 
     public function getValue(): SecondType
     {
-        $variable = new AnotherType;
+        $variable = new AnotherType();
         $assignedVariable = $variable;
     }
 }
