+++ packages/node-type-resolver/tests/PerNodeTypeResolver/VariableTypeResolver/Source/SecondType.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\VariableTypeResolver\Source;
+
+final class SecondType
+{
+
+}
