+++ packages/node-type-resolver/tests/PerNodeTypeResolver/VariableTypeResolver/VariableTypeResolverTest.php
@@ -10,8 +10,8 @@ use PHPStan\Type\ObjectType;
 use PHPStan\Type\ThisType;
 use PHPStan\Type\TypeWithClassName;
 use Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\AbstractNodeTypeResolverTest;
-use Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\VariableTypeResolver\Fixture\AnotherType;
 use Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\VariableTypeResolver\Fixture\ThisClass;
+use Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\VariableTypeResolver\Source\AnotherType;
 use ReflectionClass;
 
 /**
@@ -35,13 +35,12 @@ final class VariableTypeResolverTest extends AbstractNodeTypeResolverTest
 
     public function provideData(): Iterator
     {
-        yield [__DIR__ . '/Fixture/ThisClass.php', 0, new ThisType(new ReflectionClass(ThisClass::class))];
+        yield [__DIR__ . '/Fixture/this_class.php.inc', 0, new ThisType(new ReflectionClass(ThisClass::class))];
 
         $anotherTypeObjectType = new ObjectType(AnotherType::class);
-
-        yield [__DIR__ . '/Fixture/NewClass.php', 1, $anotherTypeObjectType];
-        yield [__DIR__ . '/Fixture/NewClass.php', 3, $anotherTypeObjectType];
-        yield [__DIR__ . '/Fixture/AssignmentClass.php', 2, $anotherTypeObjectType];
-        yield [__DIR__ . '/Fixture/ArgumentTypehint.php', 1, $anotherTypeObjectType];
+        yield [__DIR__ . '/Fixture/new_class.php.inc', 1, $anotherTypeObjectType];
+        yield [__DIR__ . '/Fixture/new_class.php.inc', 3, $anotherTypeObjectType];
+        yield [__DIR__ . '/Fixture/assignment_class.php.inc', 2, $anotherTypeObjectType];
+        yield [__DIR__ . '/Fixture/argument_typehint.php.inc', 1, $anotherTypeObjectType];
     }
 }
