+++ rules/symfony5/tests/Rector/MethodCall/FormBuilderSetDataMapperRector/Fixture/fixture.php.inc
@@ -5,7 +5,7 @@ namespace Rector\Symfony5\Tests\Rector\MethodCall\FormBuilderSetDataMapperRector
 use Symfony\Component\Form\Extension\Core\DataMapper\PropertyPathMapper;
 use Symfony\Component\Form\FormConfigBuilderInterface;
 
-class Fixture
+final class Fixture
 {
     public function run(FormConfigBuilderInterface $builder)
     {
@@ -20,7 +20,7 @@ namespace Rector\Symfony5\Tests\Rector\MethodCall\FormBuilderSetDataMapperRector
 use Symfony\Component\Form\Extension\Core\DataMapper\PropertyPathMapper;
 use Symfony\Component\Form\FormConfigBuilderInterface;
 
-class Fixture
+final class Fixture
 {
     public function run(FormConfigBuilderInterface $builder)
     {
