+++ rules/symfony5/tests/Rector/MethodCall/FormBuilderSetDataMapperRector/Fixture/skip_correct_arg.php.inc
@@ -5,7 +5,7 @@ namespace Rector\Symfony5\Tests\Rector\MethodCall\FormBuilderSetDataMapperRector
 use Symfony\Component\Form\Extension\Core\DataMapper\PropertyPathMapper;
 use Symfony\Component\Form\FormConfigBuilderInterface;
 
-class SkipCorrectArg
+final class SkipCorrectArg
 {
     public function run(FormConfigBuilderInterface $builder)
     {
