--- rules/type-declaration/tests/Rector/ClassMethod/AddReturnTypeDeclarationRector/Fixture/fixture.php.inc
@@ -1,41 +1,0 @@
-<?php
-
-namespace Rector\TypeDeclaration\Tests\Rector\ClassMethod\AddReturnTypeDeclarationRector\Fixture;
-
-class Fixture
-{
-    public function parse()
-    {
-    }
-
-    public function resolve()
-    {
-    }
-
-    public function nullable(): array
-    {
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\TypeDeclaration\Tests\Rector\ClassMethod\AddReturnTypeDeclarationRector\Fixture;
-
-class Fixture
-{
-    public function parse(): array
-    {
-    }
-
-    public function resolve(): \SomeType
-    {
-    }
-
-    public function nullable(): ?\SomeType
-    {
-    }
-}
-
-?>
