--- rules/type-declaration/tests/Rector/ClassMethod/AddReturnTypeDeclarationRector/Fixture/remove_return_type.php.inc
@@ -1,25 +1,0 @@
-<?php
-
-namespace Rector\TypeDeclaration\Tests\Rector\ClassMethod\AddReturnTypeDeclarationRector\Fixture;
-
-class RemoveReturnType
-{
-    public function clear(): array
-    {
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\TypeDeclaration\Tests\Rector\ClassMethod\AddReturnTypeDeclarationRector\Fixture;
-
-class RemoveReturnType
-{
-    public function clear()
-    {
-    }
-}
-
-?>
