+++ stubs/PHPOffice/PHPExcel/CalcEngine/PHPExcel_CalcEngine_Logger.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+if (class_exists('PHPExcel_CalcEngine_Logger')) {
+    return;
+}
+
+final class PHPExcel_CalcEngine_Logger
+{
+}
+
