+++ stubs/PHPOffice/PHPExcel/Cell/PHPExcel_Cell_DataValidation.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+if (class_exists('PHPExcel_Cell_DataValidation')) {
+    return;
+}
+
+final class PHPExcel_Cell_DataValidation
+{
+}
