+++ stubs/PHPOffice/PHPExcel/PHPExcel_Cell.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+if (class_exists('PHPExcel_Cell')) {
+    return;
+}
+
+final class PHPExcel_Cell
+{
+
+}
