+++ stubs/PHPOffice/PHPExcel/Style/Conditional.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+if (class_exists('PHPExcel_Style_Conditional')) {
+    return;
+}
+
+final class PHPExcel_Style_Conditional
+{
+}
