+++ stubs/Symfony/Component/Form/FormConfigBuilderInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symfony\Component\Form;
+
+if (interface_exists('Symfony\Component\Form\FormConfigBuilderInterface')) {
+    return;
+}
+
+interface FormConfigBuilderInterface
+{
+
+}
